+++ packages/localizations/src/el-GR.ts
@@ -271,7 +271,7 @@ export const elGR: LocalizationResource = {
       headerTitle__general: 'General',
       headerTitle__members: 'Μέλη',
       profileSection: {
-        primaryButton: undefined,
+        primaryButton: 'Ενημέρωση προφίλ',
         title: 'Organization Profile',
         uploadAction__title: 'Logo',
       },
@@ -335,7 +335,7 @@ export const elGR: LocalizationResource = {
     },
     alternativeMethods: {
       actionLink: 'Λήψη βοήθειας',
-      actionText: 'Don’t have any of these?',
+      actionText: 'Δεν έχετε κανένα από αυτά;',
       blockButton__backupCode: 'Χρήση ενός εφεδρικού κωδικού',
       blockButton__emailCode: 'Αποστολή κωδικού με email στο {{identifier}}',
       blockButton__emailLink: 'Αποστολή συνδέσμου στο {{identifier}}',
@@ -349,7 +349,8 @@ export const elGR: LocalizationResource = {
           'Εάν αντιμετωπίζετε δυσκολία στη σύνδεση στον λογαριασμό σας, στείλτε μας email και θα επικοινωνήσουμε μαζί σας για να αποκαταστήσουμε την πρόσβαση το συντομότερο δυνατόν.',
         title: 'Λήψη βοήθειας',
       },
-      subtitle: 'Facing issues? You can use any of these methods to sign in.',
+      subtitle:
+        'Αντιμετωπίζετε δυσκολίες; Μπορείτε να χρησιμοποιήσετε οποιαδήποτε από αυτές τις μεθόδους για να συνδεθείτε.',
       title: 'Χρήση μιας άλλης μεθόδου',
     },
     backupCodeMfa: {
@@ -400,10 +401,10 @@ export const elGR: LocalizationResource = {
     forgotPassword: {
       formTitle: 'Επαναφορά κωδικού πρόσβασης',
       resendButton: 'Δεν λάβατε κωδικό; Αποστολή ξανά',
-      subtitle: 'to reset your password',
-      subtitle_email: 'First, enter the code sent to your email ID',
-      subtitle_phone: 'First, enter the code sent to your phone',
-      title: 'Reset password',
+      subtitle: 'για να επαναφέρετε τον κωδικό σας',
+      subtitle_email: 'Πρώτα, εισάγετε τον κωδικό που στάλθηκε στη διεύθυνση email σας',
+      subtitle_phone: 'Πρώτα, εισάγετε τον κωδικό που στάλθηκε στο τηλέφωνό σας',
+      title: 'Επαναφορά κωδικού πρόσβασης',
     },
     forgotPasswordAlternativeMethods: {
       blockButton__resetPassword: 'Επαναφορά κωδικού πρόσβασης',
@@ -708,11 +709,11 @@ export const elGR: LocalizationResource = {
       title: 'Προσθήκη διεύθυνσης email',
       verifyTitle: 'Verify email address',
     },
-    formButtonPrimary__add: 'Add',
+    formButtonPrimary__add: 'Προσθήκη',
     formButtonPrimary__continue: 'Συνέχεια',
     formButtonPrimary__finish: 'Ολοκλήρωση',
-    formButtonPrimary__remove: 'Remove',
-    formButtonPrimary__save: 'Save',
+    formButtonPrimary__remove: 'Αφαίρεση',
+    formButtonPrimary__save: 'Αποθήκευση',
     formButtonReset: 'Ακύρωση',
     mfaPage: {
       formHint: 'Επιλέξτε μια μέθοδο για προσθήκη.',
@@ -883,7 +884,7 @@ export const elGR: LocalizationResource = {
         title: 'Αριθμοί τηλεφώνου',
       },
       profileSection: {
-        primaryButton: undefined,
+        primaryButton: 'Ενημέρωση προφίλ',
         title: 'Προφίλ',
       },
       usernameSection: {